Add files via upload
It involves a combination of data analysis, team performance assessment, and expert opinions to make an educated guess about the team most likely to win the tournament. Predictions can be made before the IPL season begins or as the tournament progresses, and they are a common topic of discussion among cricket enthusiasts and sports analysts.